+++ leechbot/downloader/__init__.py
@@ -20,6 +20,10 @@ from .telegram import TelegramDownload, media_Identifier
 from .ytdl import YTDL_Status, get_YT_Name
 from .terabox import terabox_download
 from .gallery import gallery_download, is_gallery_link, get_gallery_name
+from .gofile import gofile_download, is_gofile
+from .bunkr import bunkr_download, is_bunkr
+from .catbox import catbox_download, is_catbox
+from .streamtape import streamtape_download, is_streamtape_link
 from .manager import downloadManager, calDownSize, get_d_name
 
 __all__ = [
@@ -40,6 +44,14 @@ __all__ = [
     "gallery_download",
     "is_gallery_link",
     "get_gallery_name",
+    "gofile_download",
+    "is_gofile",
+    "bunkr_download",
+    "is_bunkr",
+    "catbox_download",
+    "is_catbox",
+    "streamtape_download",
+    "is_streamtape_link",
     "downloadManager",
     "calDownSize",
     "get_d_name",